third_party/WebKit/LayoutTests/http/tests/devtools/console/console-viewport-selection-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that console viewport handles selection properly.


Running: testSelectionSingleLineText
Selected text:<<<EOL
nsole
EOL

Running: testReversedSelectionSingleLineText
Selected text:<<<EOL
nsole
EOL

Running: testSelectionMultiLineText
Selected text:<<<EOL
ole-viewport-selection.js:13 Message #74
console-viewport-selection.js:13 Message #75
console
EOL

Running: testSimpleVisibleSelection
anchor = {item: 72, offset: 6}, head = {item: 77, offset: 6}

Running: testHalfScrollSelectionUp
anchor = {item: 72, offset: 6}, head = {item: 77, offset: 6}

Running: testHalfScrollSelectionDown
anchor = {item: 72, offset: 6}, head = {item: 77, offset: 6}

Running: testScrollSelectionAwayUp
anchor = {item: 72, offset: 6}, head = {item: 77, offset: 6}

Running: testScrollSelectionAwayDown
anchor = {item: 72, offset: 6}, head = {item: 77, offset: 6}

Running: testShiftClickSelectionOver
anchor = {item: 72, offset: 6}, head = {item: 10, offset: 0}

Running: testShiftClickSelectionBelow
anchor = {item: 72, offset: 6}, head = {item: 140, offset: 0}

Running: testRemoveSelection
anchor = null, head = null

Running: testReversedVisibleSelection
anchor = {item: 76, offset: 6}, head = {item: 71, offset: 6}

Running: testShiftClickReversedSelectionOver
anchor = {item: 76, offset: 6}, head = {item: 10, offset: 0}

Running: testShiftClickReversedSelectionBelow
anchor = {item: 76, offset: 6}, head = {item: 140, offset: 0}

Running: testZeroOffsetSelection
Selected text: world

Running: testSelectAll
Selected all 150 messages.

Running: testSelectWithNonTextNodeContainer
Selected text: console-viewport-selection.js:13 Message #1

Selected text: console-viewport-selection.js:13 Message #1

Selected text: console-viewport-selection.js:13 Message #1
